+++ backend/laravel-api/app/Http/Requests/StoreStockAdjustmentRequest.php
@@ -19,7 +19,9 @@ class StoreStockAdjustmentRequest extends FormRequest
             'remarks' => ['nullable', 'string'],
             'items' => ['required', 'array', 'min:1'],
             'items.*.item_id' => ['required', 'uuid', 'exists:items,id'],
-            'items.*.counted_qty' => ['required', 'integer', 'min:0'],
+            // Whole-number-vs-decimal enforcement happens in StockAdjustmentService via
+            // QtyCategoryValidator (needs the Item loaded, not available here).
+            'items.*.counted_qty' => ['required', 'numeric', 'min:0'],
             'items.*.reason' => ['required', 'string'],
         ];
     }
