+++ backend/laravel-api/app/Http/Requests/UpdatePurchaseReturnRequest.php
@@ -22,7 +22,7 @@ class UpdatePurchaseReturnRequest extends FormRequest
             'remarks' => ['nullable', 'string'],
             'items' => ['sometimes', 'array'],
             'items.*.purchase_invoice_item_id' => ['required_with:items', 'uuid', 'exists:purchase_invoice_items,id'],
-            'items.*.qty_returned' => ['sometimes', 'integer', 'min:0'],
+            'items.*.qty_returned' => ['sometimes', 'numeric', 'min:0'],
             'items.*.amount' => ['required_with:items', 'numeric', 'min:0'],
         ];
     }
